refactor: XYPosition is special (XY)Vector
